refacto(components)
- putted all reutilizable components accross application in a `shared` folder
