+++ src/app/components/shared/Inputs/SmartInput/index.tsx
