+++ src/app/components/shared/index.ts
@@ -1,0 +1,8 @@
+export * from './Buttons';
+export * from './Dropdowns';
+export * from './FilterItem';
+export * from './HeaderPage';
+export * from './Inputs';
+export * from './Modals';
+export * from './SmartSelect';
+export * from './Spinner';
