--- src/app/components/various/index.tsx
@@ -1,2 +1,0 @@
-export * from './FilterItem';
-export * from './HeaderPage';
